add unit test upload-view

+++ src/lib/ngx-fileupload/libs/ui/src/upload-view.ts
@@ -72,7 +72,7 @@ export class UploadViewComponent implements OnInit, OnDestroy {
         this.destroyed$.next(true);
 
         /** we handle our own storage so destroy this one */
-        if (!this.uploadStorageSet) {
+        if (!this.uploadStorageSet && this.uploadStorage) {
             this.uploadStorage.destroy();
             this.uploadStorage = null;
         }
@@ -82,8 +82,14 @@ export class UploadViewComponent implements OnInit, OnDestroy {
      * files get dropped
      */
     public dropFiles(files: File[]) {
+
+        if (!files.length) {
+            return;
+        }
+
         const uploads = this.uploadFactory.createUploadRequest(
             files, this.uploadOptions, this.validator);
+
         this.uploadStorage.add(uploads);
     }
